+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add `Object::metrics` method and `Metrics` struct
 * Update `tokio` dependency to version `1.5.0`
 * Add `post_create`, `pre_recycle` and `post_recycle` hooks
+* Add `Pool::resize` method
 
 ## v0.8.2